Urduocr/requirement.txt
asn1crypto==0.24.0
astroid==1.6.5
cffi==1.11.5
colorama==0.3.9
cryptography==2.2.2
Django==2.0.1
FormEncode==1.3.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
numpy==1.14.5
opencv-python==3.4.1.15
other==1.0.0
Pillow==5.2.0
pycparser==2.18
PyDispatcher==2.0.5
pylint==1.9.2
PyMySQL==0.9.2
pytesseract==0.2.2
pytz==2018.5
six==1.11.0
SQLObject==3.7.0
that==1.0.7
wrapt==1.10.11
zope.component==4.4.1
zope.event==4.3.0
zope.interface==4.5.0
